分页查询：根据地址搜索交易输出、根据地址搜索未花费交易输出、查询挖矿中的交易

+++ helloworld-blockchain-core/src/main/java/com/xingkaichun/helloworldblockchain/core/MinerTransactionDtoDataBase.java
@@ -25,7 +25,7 @@ public abstract class  MinerTransactionDtoDataBase {
     /**
      * 获取交易
      */
-    public abstract List<TransactionDTO> selectTransactionDtoList(BlockChainDataBase blockChainDataBase,int from, int size) throws Exception ;
+    public abstract List<TransactionDTO> selectTransactionDtoList(BlockChainDataBase blockChainDataBase,long from, long size) throws Exception ;
 
     /**
      * 删除交易
